requirements.txt
customtkinter~=5.2.0
Roboto font
pandas~=2.0.3
Pillow~=10.0.0
CTkTable~=0.8